Expose `taDevKit` in `config.system.build.jetsonDevicePkgs.taDevKit`
The TA dev kit is specific per-board and it is necessary for this to be
exposed in order for users to build trusted applications that don't get
built into the OPTEE image.

This can be used when building a TA by using the make flag
"TA_DEV_KIT_DIR=${config.system.build.jetsonDevicePkgs.taDevKit}/export-ta_arm64".